Re-add AppFigures as a data source (JEEVES-51) (#189)
* Add AppFigures with new data structure (JEEVES-51)

This has a bunch of problems still but I need to commit what I have to merge
the new data structure change

* Re-add AppFigures as a data source (JEEVES-51)

This one took some doing. A lot of functionality had to be shifted around to
avoid recursive import statements. Downloading new documents was moved into
the classes for those documents as classmethods, which felt a little weird but
it works. I also had to modify the Javascript a little bit and I had to make
a few mapping types explicit.

* Address GitHub feedback

* Address further GitHub feedback