Add support for JIRA issues (DEL-347) (#195)
* Add support for JIRA issues (DEL-347)

This change highlights the fruits of my labors on previous changes; adding a whole new document type was as easy as writing the model file for it and adding it to the identifier/document mapping. Through testing so far I've discovered that the data JIRA returns to us doesn't always match their spec, which means (for example) checking for missing fields is not consistent in this change. The checking that is present is the result of running the downloader and seeing where it crahsed. If anyone knows how to introduce checking for all input fields without making the code into a spaghetti mess I'm open to suggestions.

* Add labels field to jira documents for Hideki

* Add temporary username and API token to terraform configuration.

* Separate downloader functions from document classes

This is a rough draft of what the separated code will eventually look like but I do not have sufficient brainpower to continue on this trek today so I will polish it tomorrow

* Respond to GitHub feedback and make some miscellaneous changes

Downloaders have been moved into manager classes, which are basically downloaders but with a function to get the class of document that gets downloaded. This way we can have the identifier/document mapping become an identifier/manager mapping, and by making this mapping a class we can avoid any magic indices. Also switched the downloaders over to using Session objects, which is probably something we should have done a while ago.

* Increment major data version number

* Remove block of commented-out code

* Modify JIRA download loop to handle cases where fewer issues are downloaded than expected

Apparently JIRA can just decide to ignore your download request parameters and return 100 results instead of 1000.

* Remove dead import

Co-authored-by: Peter Chapman <peter@duolingo.com>

Co-authored-by: Peter Chapman <peter@duolingo.com>