Add explicit environment variables for general JIRA login in the API subservice (#230)
The API needs to log into Jira in one very specific circumstance that applies to both Android and iOS, so using either of those logins doesn't feel appropriate. This does mean we're using three different logins to the same service in a single module, which I would like to change later, but this fixes a bug in duplicate detection that I'm told is somewhat urgent so this will do for now.