Standardize shake-to-report metadata (JEEVES-119) (#256)
* Standardize metadata from shake-to-report where possible (JEEVES-119)

This change reorganizes some of the metadata that we attach to shake-to-report documents. Certain fields on Android and iOS provide us essentially the same information in different formats and different field names. For details about which field names turn into which standardized names, please see metadata_standardizer.py.

This change also increments the single-minor data version number and resets the double-minor data version number, because this change will change our data format and thus we will need to re-download all our records.

Finally, this change adds a field `jeeves_uid` to our document data structure, which directly contains the unique ID used to look up an individual document in a recent update. The addition of this field is not related to the other parts of this change, however it would also require us to increment the data version number and redownload all our records, which this change already needs to do, so I thought we could save on a redownload cycle by stapling these changes together.

* Add more detailed comments before code review

Also found a bug with how we're parsing fullstory URLs on iOS, which will be fixed in a separate ticket.

* Mark certain fields as keywords to bypass Elasticsearch's analyzer

* Intermediate commit while I fix the fullstory parser bug

* More intermediate commit

* Intermediate commit while I work on fixing the markupsafe import bug

* Clean up code a little

* Added comments to new functions

* Add tests