Update Duplicate Detector (JEEVES-143) (#310)
* Initial commit for updating duplicate detector (JEEVES-143)

This commit does quite a number of things. The most major revision is that Dockerfile.dev now uses Ubuntu instead of Alpine as its base image. The data versioning number has also been updated. Several methods have been shuffled around to sidestep problems with circular dependencies, and some old unused code was removed. This commit is not yet ready for production deployment, as the production Dockerfile has not been updated and documentation still needs to be worked on, and the new duplicate detector algorithm is not complete as it has no guarantees around number of returned results or same-project filtering. However, I would like to deploy this as a smoke test before Monday to give the testing instace time to populate.

* Add git lfs and set it to track .bin files

* Add files for duplicate detector model

* Remove sensitive information from Dockerfile.dev

* Update jenkins-build-dev.sh to sync duplicate detector model from S3.

* Correct bad envvar name in jenkins-build-dev.sh and remove now unused copy of duplicate detector model

* Add an ls to help debug build process

* Have jenkins-build-dev.sh explicitly create a directory to store duplicate detection model

* Exhange order of AppFigures and Jira on a temporary basis to facilitate testing of new duplicate detector

* Correct use of argument in embedding vector calculation

* Finish base implementation of new duplicate algorithm

* Small edit to correct for library implementation

* Implement pre-filtered knn query

* Revert script scoring change, add max depth search parameter to duplicate detector

* Add missing colon to make return value become list

* Cleanup various places, add comments, copy Docker changes from dev to prod

Ideally, this is the last commit before review

* Address github feedback

* Change definition order of identifier/manager mapping

Strictly according to the spec, this shouldn't change the iteration order over the mapping, however I previously changed this order for testing and I believe it behaved as I wanted, namely, it moved Jira to the head of the ordering. If that doesn't happen this time I'll implement a more formal solution.

* Prevent documents from being reported as potential duplicates of themselves

* Filter out known duplicates and clones from potential duplicates

* Add resolution field to Jira issues at Jesse's request, increment double minor data version number to reflect this

* Remove Dockerfile.dev.MOD