Fix main URL redirect (TRI-2827) (#627)
Pin the version of the `path-to-regexp` to `v1.8.0` so that it is not overwritten by the older (incompatible) version required by the dev dependency `local-cors-proxy`.

## Description

The main URL of https://jeeves.duolingo.com/ used to redirect to `/en`; now it shows a blank page with JavaScript errors in the console:

Uncaught TypeError: path_to_regexp__WEBPACK_IMPORTED_MODULE_5___default.compile is not a function

In #617, I upgraded several libraries and then had to regenerate `package-lock.json`. In doing so, the package `path-to-regexp` (`v0.1.7` required by `express` as used by `local-cors-proxy`, and also `v1.8.0` as required by `react-router`) resolved to `v0.1.7`, whereas it previously resolved to `v1.8.0`.

Per [this answer on StackOverflow](https://stackoverflow.com/a/62640965), the best fix is to pin the version of `path-to-regexp` to `v1.8.0`.