Pull log summarizations from S3 and add to ticket (#1245)
## Description

This PR enhances the log summarization feature by integrating it with the duplicate detection workflow. It adds functionality to retrieve log summaries from S3 and display them in Jira tickets as rich text in an expandable section titled "Relevant Logs".

This will enable Log summarization on JIRA tickets in Prod.

https://duolingo.atlassian.net/browse/DLAI-59083
![Screenshot 2025-05-14 at 11.14.29 AM.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/VkQoxlw0WIHC3gh1WqJH/fd54a716-f937-4f01-a7e3-2c4a39d1dbd6.png)

## Verification steps

- [x] Confirm that log summaries appear in Jira tickets for enabled features